requirements.txt
discord~=2.1.0
pytesseract~=0.3.10
requests~=2.28.1
python-dotenv~=0.21.0
Pillow~=9.3.0